+++ src/types/index.ts
@@ -13,3 +13,4 @@ export * from './message';
 export * from './stats';
 export * from './curriculum';
 export * from './notifications';
+export * from './companyNotifications';